Move apply rgb from theme to plugin
