+++ plugin/assets/src/helper/apply-rgb-value.js
